feat: main-service:
- доработан функционал по работе с комментариями. Реализован приватный контроллер для работы с комментариями пользователей

+++ main-service/src/main/java/ru/yandex/practicum/mainserver/event/comment/CommentServiceImpl.java
@@ -7,12 +7,14 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.mainserver.event.EventService;
 import ru.yandex.practicum.mainserver.event.comment.model.Comment;
+import ru.yandex.practicum.mainserver.event.model.Event;
 import ru.yandex.practicum.mainserver.exception.ConflictException;
 import ru.yandex.practicum.mainserver.exception.ObjectNotFountException;
 import ru.yandex.practicum.mainserver.request.RequestService;
 import ru.yandex.practicum.mainserver.request.model.Request;
 import ru.yandex.practicum.mainserver.status.Status;
 import ru.yandex.practicum.mainserver.user.UserService;
+import ru.yandex.practicum.mainserver.user.model.User;
 
 import java.time.LocalDateTime;
 import java.util.Collection;
@@ -36,6 +38,10 @@ public class CommentServiceImpl implements CommentService {
     @Override
     public Comment addNewComment(Comment comment, Long eventId, Long userId) {
         validateUsersRequestAndEvent(userId, eventId);
+        Event event = eventService.getEventById(eventId);
+        User user = userService.getUserById(userId);
+        comment.setEvent(event);
+        comment.setAuthor(user);
         try {
             log.info("CommentServiceImpl: addNewComment — Комментарий добавлен {}.", comment);
             return repository.save(comment);
@@ -48,7 +54,7 @@ public class CommentServiceImpl implements CommentService {
     @Override
     public Comment changeCommentByUser(Comment updComment, Long eventId, Long userId) {
         validateUsersRequestAndEvent(userId, eventId);
-        Comment comment = getCommentById(updComment.getId());
+        Comment comment = getCommentById(updComment.getId(), userId);
         comment.setText(updComment.getText());
         try {
             log.info("CommentServiceImpl: addNewComment — Комментарий изменён {}.", comment);
@@ -61,30 +67,37 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     public Collection<Comment> findAllByEventIdOrderByCreatDesc(Long eventId, Pageable pageable) {
-        return repository.findByAuthorIdOrderByCreatedDesc(eventId, pageable).toList();
+        return repository.findByEventIdOrderByCreatedDesc(eventId, pageable).toList();
     }
 
     @Override
     public Collection<Comment> findAllByEventIdOrderByCreatDesc(Long eventId) {
-        return repository.findByAuthorIdOrderByCreatedDesc(eventId);
+        return repository.findByEventIdOrderByCreatedDesc(eventId);
     }
 
     @Override
-    public void removeComment(Long commentId) {
-        getCommentById(commentId); // проверка, что комментарий с указанным id есть
+    public void removeComment(Long commentId, Long userId) {
+        Comment comment = getCommentById(commentId, userId); // проверка, что комментарий с указанным id есть
+        if (!comment.getAuthor().getId().equals(userId)) {
+            log.warn("CommentServiceImpl: removeComment — пользователь с id {} не является автором комментария c id {}", userId, commentId);
+            throw new ConflictException("Пользователь с  id " + userId + " не является автором комментария c id " + commentId);
+        }
         log.info("CommentServiceImpl: removeComment — Комментарий с указанным id {} удалён", commentId);
         repository.deleteById(commentId);
 
     }
 
     @Override
-    public Comment getCommentById(Long commentId) {
+    public Comment getCommentById(Long commentId, Long userId) {
         Optional<Comment> comment = repository.findById(commentId);
         comment.orElseThrow(() -> {
             log.warn("CommentServiceImpl: getCommentById — Комментария с указанным id {} нет", commentId);
-            return new ObjectNotFountException("Комментария с указанным id " + commentId + " нет");
+            throw new ObjectNotFountException("Комментария с указанным id " + commentId + " нет");
         });
-
+        if (!comment.get().getAuthor().getId().equals(userId)) {
+            log.warn("CommentServiceImpl: removeComment — пользователь с id {} не является автором комментария c id {}", userId, commentId);
+            throw new ConflictException("Пользователь с  id " + userId + " не является автором комментария c id " + commentId);
+        }
         log.warn("CommentServiceImpl: getCommentById — Комментарий с указанным id {} получен", commentId);
         return comment.get();
     }
